+++ app/src/main/java/com/example/myuniversityclient/data/repository/main/ITLinksService.kt
@@ -3,5 +3,5 @@ package com.example.myuniversityclient.data.repository.main
 import com.example.myuniversityclient.data.models.ITServicesList
 
 interface ITLinksService {
-    fun getITServices(onResult: (Result<ITServicesList?>) -> Unit)
+    fun getITServices(onResult: (Result<ITServicesList>) -> Unit)
 }
