+++ app/src/main/java/com/example/myuniversityclient/data/repository/main/MainRepository.kt
@@ -2,7 +2,6 @@ package com.example.myuniversityclient.data.repository.main
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import com.example.myuniversityclient.data.models.ITServicesList
 import com.example.myuniversityclient.data.models.ShortUserInfo
 import javax.inject.Inject
 
@@ -10,15 +9,13 @@ import javax.inject.Inject
  * A repository used in [com.example.myuniversityclient.MainActivity].
  */
 class MainRepository @Inject constructor(
-    private val service: MainService,
-    private val itLinksService: ITLinksService
+    private val service: MainService
 ) {
     /**
      * Last returned user info [LiveData].
      * Stored to later have the ability to send further updates to observers.
      */
     private var userInfo: MutableLiveData<Result<ShortUserInfo?>>? = null
-    private var itServices: MutableLiveData<Result<ITServicesList?>>? = null
 
     /**
      * Provides observable [LiveData] with requested user info to observers.
@@ -42,16 +39,5 @@ class MainRepository @Inject constructor(
     fun logout() {
         userInfo?.value = null
         service.logout()
-    }
-
-    fun getITServicesList(): LiveData<Result<ITServicesList?>> {
-        val liveData = MutableLiveData<Result<ITServicesList?>>()
-        itServices = liveData
-
-        itLinksService.getITServices {
-            liveData.value = it
-        }
-
-        return liveData
     }
 }
