+++ app/src/main/java/com/example/myuniversityclient/domain/ITServicesFragmentViewModel.kt
@@ -4,13 +4,14 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.Transformations
 import androidx.lifecycle.ViewModel
 import com.example.myuniversityclient.data.models.ITServicesList
+import com.example.myuniversityclient.data.repository.main.ITLinksRepository
 import com.example.myuniversityclient.data.repository.main.MainRepository
 import com.example.myuniversityclient.ui.models.ITServicesListModel
 import com.example.myuniversityclient.ui.models.ShortUserInfoModel
 import javax.inject.Inject
 
 class ITServicesFragmentViewModel @Inject constructor(
-    private val repository: MainRepository
+    private val repository: ITLinksRepository
 ): ViewModel() {
 
     fun getITServicesList(): LiveData<Result<ITServicesList?>> {
